requirement.txt
Flask==2.3.2
openai==0.27.8
requests==2.31.0
Pillow==10.0.0
pyttsx3==2.90
speechrecognition==3.8.1